Use React 16.3+ Context API
To keep in line with React’s context component naming scheme (Provider and Consumer), this change introduces a new `<ActionCableConsumer />` component.

This new component has the same API as the old `<ActionCable />` component. The old component can still be used—it wraps the new component—but will output a deprecation warning in the console.

The `<ActionCableProvider />` component’s API is unchanged.

As this bumps the required React version from 15.6 to 16.3, **this is a breaking change.**

If you’re using React <16.3, you will be unable to use this package until you upgrade.

If you’re already using React >16.3, you shouldn’t have to change anything, but you *will* get a deprecation warning in the console for using the `<ActionCable />` component.